% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcTSP4.R
\name{extractArtemisFormatFromTSP4}
\alias{extractArtemisFormatFromTSP4}
\title{Extract plot list for Artemis simulation}
\usage{
extractArtemisFormatFromTSP4(QcTSP4Data, plots, format = "WebAPI")
}
\arguments{
\item{QcTSP4Data}{the database that is retrieved through the restoreQcTSP4Data function}

\item{plots}{a vector of integers standing for the plot id to be considered}

\item{format}{a character string identifying the intended use, either "WebAPI" or "RPackage"}
}
\value{
a data.frame object formatted for Capsis Web API
}
\description{
Extract plot list for Artemis simulation
}
\examples{
\dontrun{
 extractArtemisFormatFromTSP4(QcTSP4Data=Data, plots= c(700412604, 504656201), format="WebAPI")}

}
