% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcTSP4.R
\name{extractNatura2014FormatFromTSP4ForMetaModelling}
\alias{extractNatura2014FormatFromTSP4ForMetaModelling}
\title{Extract plot list for Natura simulation}
\usage{
extractNatura2014FormatFromTSP4ForMetaModelling(QcTSP4Data, stratumPlots)
}
\arguments{
\item{QcTSP4Data}{the database that is retrieved through the restoreQcTSP4Data function}

\item{stratumPlots}{a dataframe with a column "stratum" as a grouping variable as a character field and a column "plots" for plots number as an integer}
}
\value{
a list of two dataframes where the first is a tree list and the second is a list of study trees for model Natura on Capsis Web API
}
\description{
Extract plot list for Natura simulation
}
